made some refactoring of feedback related files. Feedback1 doesn't exist, there's feedback and there's feedbackbackup
